requirements.txt
RPi.GPIO==0.6.4
Adafruit-CharLCD=1.1.1
flask
flask_socketio
wiringpi